reducing to primary happy paths
